+++ src/main/java/net/jakobnielsen/imagga/crop_slice/convert/DivisionRegionConverter.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2013 Jakob Vad Nielsen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package net.jakobnielsen.imagga.crop_slice.convert;
 
 import net.jakobnielsen.imagga.convert.Converter;
@@ -11,7 +27,7 @@ import org.json.simple.JSONValue;
 import java.util.ArrayList;
 import java.util.List;
 
-import static net.jakobnielsen.imagga.convert.ConverterTools.getLong;
+import static net.jakobnielsen.imagga.convert.ConverterTools.getInteger;
 import static net.jakobnielsen.imagga.convert.ConverterTools.getString;
 
 public class DivisionRegionConverter implements Converter<List<DivisionRegion>> {
@@ -52,10 +68,10 @@ public class DivisionRegionConverter implements Converter<List<DivisionRegion>> 
                         JSONObject regionObject = (JSONObject) aRegionsArray;
 
                         regions.add(new Region(
-                                getLong("x1", regionObject),
-                                getLong("y1", regionObject),
-                                getLong("x2", regionObject),
-                                getLong("y2", regionObject)
+                                getInteger("x1", regionObject),
+                                getInteger("y1", regionObject),
+                                getInteger("x2", regionObject),
+                                getInteger("y2", regionObject)
                         ));
                     }
                     divisionRegions.add(new DivisionRegion(url, regions));
